day11/day11_test.go
package main

import (
	"bytes"
	"testing"
)

var part1Test = `
125 17
`

func TestDay11(t *testing.T) {
	day11 := Day11{}
	t.Run("part 1", func(t *testing.T) {
		input := bytes.NewBufferString(part1Test)

		expected := 55312
		actual := day11.Part1(input)

		if expected != actual {
			t.Fatalf("ERROR: expected %d actual %d\n", expected, actual)
		}
	})

	t.Run("part 2", func(t *testing.T) {
		input := bytes.NewBufferString(part1Test)

		expected := 65601038650482
		actual := day11.Part2(input)

		if expected != actual {
			t.Fatalf("ERROR: expected %d actual %d\n", expected, actual)
		}
	})
}
